change REVERSE functionality of drivetrainSubsystem to avoid accessing a member variable
To avoid accessing variables that should
ideally be private, I wrote methods to get
and set the isReversed boolean.

The InputReverse command was rewritten as an InstantCommand, which executes once and only once. This prevents it from overlapping with the functionality of TeleopJoystick.

Documentation was rewrritten to avoid referencing button names (possibly causing confusion for future programmers).

-- Alex